Move dockerComposeProcessBuilder method to utils class
